Fixed make check issues
