+++ core/design-system/src/main/kotlin/es/pedrazamiguez/splittrip/core/designsystem/presentation/component/chart/DonutChart.kt
@@ -1,6 +1,8 @@
 package es.pedrazamiguez.splittrip.core.designsystem.presentation.component.chart
 
+import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.aspectRatio
@@ -38,9 +40,9 @@ fun DonutChart(
 
     val total = remember(data) { data.sumOf { it.value.toDouble() }.toFloat().coerceAtLeast(1f) }
 
-    val animationSpec = androidx.compose.animation.core.tween<Float>(
+    val animationSpec = tween<Float>(
         durationMillis = ANIMATION_DURATION_MS,
-        easing = androidx.compose.animation.core.FastOutSlowInEasing
+        easing = FastOutSlowInEasing
     )
 
     val animatedSweepAngles = data.mapIndexed { index, item ->
