+++ features/balances/src/main/kotlin/es/pedrazamiguez/splittrip/features/balance/presentation/component/PocketBalanceStatsRow.kt
@@ -1,17 +1,11 @@
 package es.pedrazamiguez.splittrip.features.balance.presentation.component
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -19,12 +13,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
 import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
 import es.pedrazamiguez.splittrip.features.balance.R
-import es.pedrazamiguez.splittrip.features.balance.presentation.model.CategorySpendingUiModel
 import es.pedrazamiguez.splittrip.features.balance.presentation.model.GroupPocketBalanceUiModel
-import kotlin.math.roundToInt
 
 @Composable
 internal fun PocketBalanceStatsRow(
@@ -71,52 +62,5 @@ internal fun PocketBalanceStatsRow(
                 onShowExtrasBreakdown = onShowExtrasBreakdown
             )
         }
-    }
-}
-
-@Composable
-fun CategorySpendingItemRow(
-    item: CategorySpendingUiModel,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Color indicator
-        Box(
-            modifier = Modifier
-                .size(16.dp)
-                .background(item.color, RoundedCornerShape(4.dp))
-        )
-
-        Spacer(modifier = Modifier.width(MaterialTheme.spacing.Medium))
-
-        // Icon
-        Icon(
-            imageVector = item.categoryIcon,
-            contentDescription = null,
-            modifier = Modifier.size(24.dp),
-            tint = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-
-        Spacer(modifier = Modifier.width(MaterialTheme.spacing.Medium))
-
-        // Category name and percentage
-        val percentage = (item.progress * 100).roundToInt()
-        Text(
-            text = "${item.categoryName} ($percentage%)",
-            style = MaterialTheme.typography.bodyMedium,
-            fontWeight = FontWeight.Medium,
-            modifier = Modifier.weight(1f)
-        )
-
-        // Amount
-        Text(
-            text = item.formattedAmount,
-            style = MaterialTheme.typography.titleSmall,
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colorScheme.onSurface
-        )
     }
 }
